+++ VoiceCommand/VoiceCommand.py
@@ -1,12 +1,6 @@
-import sys
 import os
-import setproctitle
 import random
 import webbrowser
-from pycaw.pycaw import AudioUtilities, IAudioEndpointVolume
-from ctypes import cast, POINTER
-from comtypes import CLSCTX_ALL
-import json
 import time
 import logging
 from datetime import datetime
@@ -20,64 +14,34 @@ from selenium.webdriver.support import expected_conditions as EC
 from webdriver_manager.chrome import ChromeDriverManager
 import warnings
 import gc
-import psutil
 import whisper
 import speech_recognition as sr
 import torch
 import pvporcupine
 import pyaudio
 import struct
-import ctypes
 import threading
 from queue import Queue
 import time
 import requests
 from datetime import datetime, timedelta
 import math
-from geopy.geocoders import Nominatim
 from datetime import datetime, timedelta
 from PySide6.QtWidgets import (
     QApplication,
-    QSystemTrayIcon,
-    QMenu,
-    QMainWindow,
-    QLabel,
-    QWidget,
-    QPushButton,
-    QComboBox,
-)
-from PySide6.QtGui import (
-    QIcon,
-    QAction,
-    QPainter,
-    QPixmap,
-    QImage,
-    QFont,
-    QFontDatabase,
-    QColor,
-    QFontMetrics,
 )
 from PySide6.QtCore import (
     QThread,
     Signal,
-    QTimer,
     Qt,
-    QPoint,
-    QPropertyAnimation,
-    QEasingCurve,
     QMetaObject,
     Qt,
     Q_ARG,
-    QRect,
-    Slot,
-    QBuffer,
 )
 from melo.api import TTS
 from pydub import AudioSegment
 from pydub.playback import play
-from ai_assistant import get_ai_assistant
-from collections import OrderedDict, deque
-
+from Config import config
 
 # 전역 변수 선언
 tts_model = None
@@ -88,161 +52,8 @@ icon_path = None
 volume = None
 active_timer = None
 
-try:
-    ctypes.windll.kernel32.SetConsoleTitleW("Ari Voice Command")
-except:
-    pass
-
-warnings.filterwarnings("ignore", category=FutureWarning)
-os.environ["SDL_VIDEODRIVER"] = "dummy"
-
-# MeCab 경로 지정
-mecab_path = r"C:\Program Files\MeCab\bin"
-if mecab_path not in os.environ["PATH"]:
-    os.environ["PATH"] = mecab_path + os.pathsep + os.environ["PATH"]
-
-# MeCab 사전 경로 지정
-dicdir = r"C:\Program Files\MeCab\dic\ipadic"
-os.environ["MECAB_DICDIR"] = dicdir
-
-icon_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "icon.ico")
-if not os.path.exists(icon_path):
-    print(f"경고: 아이콘 파일을 찾을 수 없습니다: {icon_path}")
-    icon_path = None  # 아이콘 파일이 없을 경우 None으로
-
-
-# 설정 파일에서 API 키 로드
-def load_config():
-    with open("config.json", "r") as f:
-        return json.load(f)
-
-
-config = load_config()
-
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 device = "cuda" if torch.cuda.is_available() else "cpu"
-
-
-# 볼륨 제어를 위한 설정
-devices = AudioUtilities.GetSpeakers()
-interface = devices.Activate(IAudioEndpointVolume._iid_, CLSCTX_ALL, None)
-volume = cast(interface, POINTER(IAudioEndpointVolume))
-
-
-# 전역 변수로 LRU 캐시 크기 설정
-IMAGE_CACHE_SIZE = 20
-
-
-# LRU 캐시 구현
-class LRUCache:
-    def __init__(self, capacity=10):
-        self.cache = OrderedDict()
-        self.capacity = capacity
-
-    def get(self, key):
-        if key not in self.cache:
-            return None
-        self.cache.move_to_end(key)
-        return self.cache[key]
-
-    def put(self, key, value):
-        if key in self.cache:
-            self.cache.move_to_end(key)
-        elif len(self.cache) >= self.capacity:
-            self.cache.popitem(last=False)
-        self.cache[key] = value
-
-
-# 이미지 로딩 및 캐싱 함수
-def load_and_cache_image(path, cache):
-    cached_image = cache.get(path)
-    if cached_image is not None:
-        return cached_image
-
-    image = QImage(path)
-    scaled_image = image.scaled(
-        image.width() * 1.5,
-        image.height() * 1.5,
-        Qt.KeepAspectRatio,
-        Qt.SmoothTransformation,
-    )
-
-    # 이미지 압축
-    buffer = QBuffer()
-    buffer.open(QBuffer.WriteOnly)
-    scaled_image.save(buffer, "PNG", quality=70)  # 압축률 조정 (0-100)
-    compressed_image = QImage()
-    compressed_image.loadFromData(buffer.data(), "PNG")
-
-    cache.put(path, compressed_image)
-    return compressed_image
-
-
-# 로그 설정
-def setup_logging():
-    log_dir = "logs"
-    if not os.path.exists(log_dir):
-        os.makedirs(log_dir)
-
-    current_time = datetime.now().strftime("%Y%m%d_%H%M%S")
-    log_file = os.path.join(log_dir, f"ari_log_{current_time}.log")
-
-    logging.basicConfig(
-        level=logging.INFO,
-        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-        handlers=[
-            logging.FileHandler(log_file, encoding="utf-8"),
-            logging.StreamHandler(sys.stdout),  # stdout으로 변경
-        ],
-    )
-
-
-# 리소스 모니터링 스레드
-class ResourceMonitor(QThread):
-    gc_needed = Signal()
-
-    def __init__(self, check_interval=5):
-        super().__init__()
-        self.check_interval = check_interval
-        self.running = True
-        self.process = psutil.Process()
-        self.memory_history = deque(maxlen=10)  # 최근 10개의 메모리 사용량 기록
-        self.last_gc_time = time.time()
-
-    def get_memory_info(self):
-        try:
-            mem_info = self.process.memory_info()
-            return mem_info.private / (1024 * 1024)  # Private Working Set in MB
-        except:
-            return 0
-
-    def run(self):
-        while self.running:
-            current_memory = self.get_memory_info()
-            self.memory_history.append(current_memory)
-
-            if len(self.memory_history) >= 5:  # 최소 5개의 샘플이 있을 때
-                avg_memory = sum(self.memory_history) / len(self.memory_history)
-                if current_memory > avg_memory * 1.5 and time.time() - self.last_gc_time > 60:
-                    # 현재 메모리가 평균의 1.5배 이상이고, 마지막 GC로부터 1분 이상 지났을 때
-                    self.gc_needed.emit()
-                    self.last_gc_time = time.time()
-
-            time.sleep(self.check_interval)
-
-    def stop(self):
-        self.running = False
-
-
-# 가비지 컬렉션 실행
-def perform_gc():
-    logging.info("가비지 컬렉션 실행 중...")
-    gc.collect()
-    process = psutil.Process()
-    mem_info = process.memory_info()
-    logging.info(f"가비지 컬렉션 완료. 메모리 사용량:")
-    logging.info(f"  Private Working Set: {mem_info.private / (1024 * 1024):.2f} MB")
-    logging.info(f"  RSS: {mem_info.rss / (1024 * 1024):.2f} MB")
 
 # 모델 로딩 스레드
 class ModelLoadingThread(QThread):
@@ -301,7 +112,7 @@ def text_to_speech(text):
         # TTS 재생 전에 음성 인식을 비활성화
         app = QApplication.instance()
         main_window = next(
-            (w for w in app.topLevelWidgets() if isinstance(w, MainWindow)), None
+            (w for w in app.topLevelWidgets() if w.objectName() == "MainWindow"), None
         )
         if (
             main_window
@@ -311,14 +122,8 @@ def text_to_speech(text):
             main_window.voice_thread.is_tts_playing = True
 
         # 말풍선 표시
-        if main_window and hasattr(main_window, "tray_icon"):
-            for character in main_window.tray_icon.character_widgets:
-                QMetaObject.invokeMethod(
-                    character,
-                    "show_speech_bubble",
-                    Qt.QueuedConnection,
-                    Q_ARG(str, text),
-                )
+        if main_window:
+            main_window.show_speech_bubble(text)
 
         # pydub를 사용하여 오디오 재생
         sound = AudioSegment.from_file(output_path)
@@ -334,9 +139,8 @@ def text_to_speech(text):
             main_window.voice_thread.is_tts_playing = False
 
         # 말풍선 숨기기
-        if main_window and hasattr(main_window, "tray_icon"):
-            for character in main_window.tray_icon.character_widgets:
-                character.hide_speech_bubble()
+        if main_window:
+            main_window.hide_speech_bubble()
 
     except Exception as e:
         logging.error(f"TTS 처리 중 오류 발생: {str(e)}")
@@ -665,365 +469,6 @@ def set_shutdown_timer_from_command(command):
         text_to_speech("종료 시간을 정확히 말씀해 주세요.")
 
 
-class CharacterWidget(QWidget):
-    toggle_voice_recognition = Signal()
-    exit_signal = Signal()
-    set_listening_state_signal = Signal(bool)
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.is_listening = False
-        self.setWindowFlags(Qt.Tool | Qt.FramelessWindowHint | Qt.WindowStaysOnTopHint)
-        self.setAttribute(Qt.WA_TranslucentBackground)
-        self.setAttribute(Qt.WA_ShowWithoutActivating)
-
-        self.initUI()
-        self.offset = QPoint()
-        self.animation = None
-        self.context_menu = None
-        self.create_context_menu()
-        self.is_moving = False
-        self.is_dragging = False
-        self.current_frame = 0
-        self.frame_timer = QTimer(self)
-        self.frame_timer.timeout.connect(self.next_frame)
-        self.last_pos = self.pos()
-
-        self.move_timer = QTimer(self)
-        self.move_timer.timeout.connect(self.start_random_move)
-
-        self.set_listening_state_signal.connect(self._set_listening_state)
-        self.facing_right = False
-
-        self.actions = ["idle", "move", "sit"]
-        self.current_action = "idle"
-        self.action_timer = QTimer(self)
-        self.action_timer.timeout.connect(self.perform_random_action)
-        self.action_duration_timer = QTimer(self)
-        self.action_duration_timer.timeout.connect(self.end_current_action)
-
-        self.image_cache = LRUCache(20)
-        self.load_images()
-        self.start_auto_move()
-
-        self.falling = False
-        self.fall_timer = QTimer(self)
-        self.fall_timer.timeout.connect(self.fall)
-
-        self.fall_frame = 0
-
-        self.interaction_timer = QTimer(self)
-        self.interaction_timer.timeout.connect(self.interact_with_others)
-        self.interaction_timer.start(5000)  # 5초마다 상호작용 시도
-
-    def initUI(self):
-        self.setGeometry(100, 100, 100, 100)
-
-    def load_images(self):
-        image_folder = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)), "images"
-        )
-        self.image_cache = LRUCache(50)  # 캐시 크기 제한
-        self.image_sets = {
-            "idle": [
-                load_and_cache_image(
-                    os.path.join(image_folder, f"idle{i}.png"), self.image_cache
-                )
-                for i in range(1, 9)
-            ],
-            "walk": [
-                load_and_cache_image(
-                    os.path.join(image_folder, f"walk{i}.png"), self.image_cache
-                )
-                for i in range(1, 10)
-            ],
-            "sit": [
-                load_and_cache_image(
-                    os.path.join(image_folder, f"sit{i}.png"), self.image_cache
-                )
-                for i in range(1, 10)
-            ],
-        }
-        self.current_image = self.image_sets["idle"][0]
-        self.resize(QPixmap.fromImage(self.current_image).size())
-
-    def load_and_cache_image(self, path):
-        cached_image = self.image_cache.get(path)
-        if cached_image is not None:
-            return cached_image
-
-        image = QImage(path)
-        scaled_image = image.scaled(
-            image.width() * 1.5,
-            image.height() * 1.5,
-            Qt.KeepAspectRatio,
-            Qt.SmoothTransformation,
-        )
-
-        # 이미지 압축
-        buffer = QBuffer()
-        buffer.open(QBuffer.WriteOnly)
-        scaled_image.save(buffer, "PNG", quality=70)  # 압축률 조정 (0-100)
-        compressed_image = QImage()
-        compressed_image.loadFromData(buffer.data(), "PNG")
-
-        self.image_cache.put(path, compressed_image)
-        return compressed_image
-
-    def get_image_set(self, action):
-        if action not in self.image_sets:
-            image_folder = os.path.join(
-                os.path.dirname(os.path.abspath(__file__)), "images"
-            )
-            if action == "drag":
-                self.image_sets[action] = [
-                    self.load_and_cache_image(
-                        os.path.join(image_folder, f"drag{i}.png")
-                    )
-                    for i in range(1, 9)
-                ]
-            elif action == "fall":
-                self.image_sets[action] = [
-                    self.load_and_cache_image(
-                        os.path.join(image_folder, f"fall{i}.png")
-                    )
-                    for i in range(1, 9)
-                ]
-        return self.image_sets[action]
-
-    def start_auto_move(self):
-        if not self.is_listening:
-            self.move_timer.start(random.randint(3000, 10000))
-
-    def stop_auto_move(self):
-        self.move_timer.stop()
-        if self.animation:
-            self.animation.stop()
-        self.is_moving = False
-        self.frame_timer.stop()
-
-    def start_random_move(self):
-        if not self.is_dragging:
-            self.animate()
-
-    def animate(self):
-        if (
-            self.animation is None
-            or self.animation.state() == QPropertyAnimation.Stopped
-        ):
-            self.animation = QPropertyAnimation(self, b"pos")
-            self.animation.setDuration(3000)
-            start_pos = self.pos()
-            self.animation.setStartValue(start_pos)
-
-            screen = QApplication.primaryScreen().geometry()
-            max_distance = 100
-            new_x = max(
-                0,
-                min(
-                    start_pos.x() + random.randint(-max_distance, max_distance),
-                    screen.width() - self.width(),
-                ),
-            )
-            new_y = max(
-                0,
-                min(
-                    start_pos.y() + random.randint(-max_distance, max_distance),
-                    screen.height() - self.height(),
-                ),
-            )
-
-            self.animation.setEndValue(QPoint(new_x, new_y))
-            self.animation.setEasingCurve(QEasingCurve.InOutQuad)
-            self.animation.finished.connect(self.animationFinished)
-            self.animation.start()
-            self.current_action = "move"
-            self.is_moving = True
-            self.frame_timer.start(100)
-            self.facing_right = new_x > start_pos.x()
-
-    def mousePressEvent(self, event):
-        if event.button() == Qt.LeftButton:
-            self.offset = event.pos()
-            self.is_dragging = True
-            self.stop_auto_move()
-            self.current_image = self.drag_images[0]
-            self.update()
-        elif event.button() == Qt.RightButton:
-            self.show_context_menu(event.pos())
-
-    def mouseMoveEvent(self, event):
-        if event.buttons() & Qt.LeftButton:
-            new_pos = self.mapToParent(event.pos() - self.offset)
-            self.move(new_pos)
-            self.update_drag_image(new_pos)
-
-    def mouseReleaseEvent(self, event):
-        if event.button() == Qt.LeftButton:
-            self.is_dragging = False
-            self.current_image = self.idle_image
-            self.update()
-            self.start_auto_move()
-
-    def mouseDoubleClickEvent(self, event):
-        if event.button() == Qt.LeftButton:
-            self.start_fall()
-
-    def update_drag_image(self, new_pos):
-        if new_pos.x() < self.last_pos.x():
-            self.current_image = self.drag_images[2]
-            self.facing_right = False
-        elif new_pos.x() > self.last_pos.x():
-            self.current_image = self.drag_images[1]
-            self.facing_right = True
-        else:
-            self.current_image = self.drag_images[0]
-        self.last_pos = new_pos
-        self.update()
-
-    def next_frame(self):
-        if self.is_moving:
-            self.current_frame = (self.current_frame + 1) % len(self.move_images)
-            self.current_image = self.move_images[self.current_frame]
-            self.update()
-
-    def paintEvent(self, event):
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.Antialiasing)
-
-        if self.falling:
-            painter.drawPixmap(self.rect(), QPixmap.fromImage(self.current_image))
-        elif self.facing_right and not self.is_dragging:
-            flipped_image = self.current_image.mirrored(True, False)
-            painter.drawPixmap(self.rect(), QPixmap.fromImage(flipped_image))
-        else:
-            painter.drawPixmap(self.rect(), QPixmap.fromImage(self.current_image))
-
-    def create_context_menu(self):
-        self.context_menu = QMenu(self)
-        sit_action = self.context_menu.addAction("앉기")
-        sit_action.triggered.connect(self.sit)
-        idle_action = self.context_menu.addAction("기본 상태")
-        idle_action.triggered.connect(self.idle)
-        fall_action = self.context_menu.addAction("떨어지기")
-        fall_action.triggered.connect(self.start_fall)
-        exit_action = self.context_menu.addAction("제거")
-        exit_action.triggered.connect(self.close)
-
-    def show_context_menu(self, pos):
-        self.context_menu.exec_(self.mapToGlobal(pos))
-
-    def closeEvent(self, event):
-        self.exit_signal.emit()
-        super().closeEvent(event)
-
-    def sit(self):
-        self.stop_auto_move()
-        self.current_action = "sit"
-        self.current_image = self.sit_image
-        self.update()
-
-    def idle(self):
-        self.current_action = "idle"
-        self.current_image = self.idle_image
-        self.update()
-        self.start_auto_move()
-
-    def animationFinished(self):
-        self.is_moving = False
-        self.current_image = self.idle_image
-        self.frame_timer.stop()
-        self.update()
-
-    def set_listening_state(self, is_listening):
-        self.set_listening_state_signal.emit(is_listening)
-
-    def _set_listening_state(self, is_listening):
-        self.is_listening = is_listening
-        if is_listening:
-            self.current_image = self.listen_image
-            self.stop_auto_move()
-            self.action_timer.stop()
-            self.action_duration_timer.stop()
-        else:
-            self.current_image = self.idle_image
-            self.start_auto_move()
-            self.action_timer.start(random.randint(15000, 45000))
-        self.update()
-
-    def perform_random_action(self):
-        if not self.is_listening and not self.is_dragging:
-            action = random.choice(["idle", "move", "sit"])
-            if action == "move":
-                self.start_random_move()
-            elif action == "sit":
-                self.sit()
-                duration = random.randint(3000, 8000)
-                self.action_duration_timer.start(duration)
-            else:
-                self.idle()
-
-        self.action_timer.start(random.randint(15000, 45000))
-
-    def end_current_action(self):
-        self.action_duration_timer.stop()
-        self.idle()
-
-    def start_fall(self):
-        if not self.falling:
-            self.falling = True
-            self.stop_auto_move()
-            self.fall_frame = 0
-            self.fall_timer.start(50)
-
-    def fall(self):
-        if not self.falling:
-            return
-
-        screen = QApplication.primaryScreen().geometry()
-        current_pos = self.pos()
-        new_pos = QPoint(current_pos.x(), current_pos.y() + 10)
-
-        if new_pos.y() + self.height() > screen.height():
-            self.falling = False
-            self.fall_timer.stop()
-            new_pos.setY(screen.height() - self.height())
-            self.current_image = self.idle_image
-            self.start_auto_move()  # 떨어진 후 다시 자동 이동 시작
-        else:
-            self.fall_frame = (self.fall_frame + 1) % len(self.fall_images)
-            self.current_image = self.fall_images[self.fall_frame]
-
-        self.move(new_pos)
-        self.update()
-
-    def interact_with_others(self):
-        if self.parent() and hasattr(self.parent(), "character_widgets"):
-            others = [char for char in self.parent().character_widgets if char != self]
-            if others:
-                target = random.choice(others)
-                self.move_towards(target.pos())
-
-    def move_towards(self, target_pos):
-        if self.falling or self.is_dragging:
-            return
-
-        current_pos = self.pos()
-        dx = target_pos.x() - current_pos.x()
-        dy = target_pos.y() - current_pos.y()
-        distance = (dx**2 + dy**2) ** 0.5
-
-        if distance > 5:
-            speed = 5
-            ratio = speed / distance
-            new_x = current_pos.x() + dx * ratio
-            new_y = current_pos.y() + dy * ratio
-            self.move(QPoint(int(new_x), int(new_y)))
-            self.current_image = self.move_images[self.current_frame]
-            self.current_frame = (self.current_frame + 1) % len(self.move_images)
-            self.update()
-
-
 class VoiceCommandThread(QThread):
     command_signal = Signal(str)
     finished = Signal()
@@ -1081,160 +526,6 @@ class VoiceCommandThread(QThread):
             self.start_listening_signal.emit()
         self.is_listening = not self.is_listening
         logging.info(f"음성 인식 {'활성화' if self.is_listening else '비활성화'}")
-
-
-class SystemTrayIcon(QSystemTrayIcon):
-    def __init__(self, icon, parent=None):
-        super(SystemTrayIcon, self).__init__(icon, parent)
-        self.setToolTip(f"Ari Voice Command")
-
-        self.menu = QMenu(parent)
-        self.setContextMenu(self.menu)
-
-        self.character_widgets = []
-        self.max_characters = 5
-
-        self.activated.connect(self.on_tray_icon_activated)
-
-        self.add_character_action = self.menu.addAction("캐릭터 추가")
-        self.add_character_action.triggered.connect(self.add_character)
-
-        self.remove_character_action = self.menu.addAction("캐릭터 제거")
-        self.remove_character_action.triggered.connect(self.remove_character)
-        self.remove_character_action.setEnabled(False)
-
-        self.exit_action = self.menu.addAction("종료")
-        self.exit_action.triggered.connect(self.exit)
-
-        self.animation_timer = QTimer(self)
-        self.animation_timer.timeout.connect(self.start_random_move)
-        self.animation_timer.start(random.randint(3000, 10000))
-
-        self.voice_thread = None  # voice_thread 속성 추가
-
-    def set_voice_thread(self, thread):
-        self.voice_thread = thread
-
-    def exit(self):
-        if self.voice_thread:
-            self.voice_thread.stop()
-            self.voice_thread.wait()
-        for character in self.character_widgets:
-            character.close()
-        self.character_widgets.clear()
-        QApplication.instance().quit()
-
-    def on_model_loaded(self):
-        # 이 메서드는 더 이상 필요하지 않지만, 오류를 방지하기 위해 빈 메서드로 유지합니다.
-        pass
-
-    def create_menu(self):
-        menu = QMenu()
-        self.add_character_action = QAction("캐릭터 추가", self)
-        self.add_character_action.triggered.connect(self.add_character)
-        menu.addAction(self.add_character_action)
-
-        self.remove_character_action = QAction("캐릭터 제거", self)
-        self.remove_character_action.triggered.connect(self.remove_character)
-        self.remove_character_action.setEnabled(False)
-        menu.addAction(self.remove_character_action)
-
-        exit_action = menu.addAction("종료")
-        exit_action.triggered.connect(self.exit_with_farewell)
-        self.setContextMenu(menu)
-
-    def handle_command(self, command):
-        logging.info(f"받은 명령: {command}")
-        if "종료" in command:
-            logging.info("프로그램을 종료합니다.")
-            self.exit_with_farewell()
-        else:
-            if self.ai_assistant is not None:
-                try:
-                    response = self.ai_assistant.process_query(command)
-                    if response:
-                        logging.info(f"AI 응답: {response}")
-                        text_to_speech(response)
-                    else:
-                        logging.error("AI 응답이 비어 있습니다.")
-                        text_to_speech(
-                            "죄송합니다. 응답을 생성하는 데 문제가 있었습니다."
-                        )
-                except Exception as e:
-                    logging.error(f"AI 응답 처리 중 오류 발생: {str(e)}")
-                    text_to_speech(
-                        "죄송합니다. 응답을 처리하는 데 문제가 발생했습니다."
-                    )
-            else:
-                logging.error("AI 어시스턴트가 초기화되지 않았습니다.")
-                text_to_speech("죄송합니다. AI 어시스턴트를 사용할 수 없습니다.")
-        for character in self.character_widgets:
-            character.set_listening_state(False)
-        self.animation_timer.start(random.randint(3000, 10000))
-
-    def exit_with_farewell(self):
-        farewells = ["안녕히 가세요.", "아리를 종료합니다."]
-        farewell = random.choice(farewells)
-        text_to_speech(farewell)
-        logging.info("프로그램을 종료합니다.")
-        for character in self.character_widgets:
-            character.close()
-        self.character_widgets.clear()
-        QTimer.singleShot(2000, self.exit)
-
-    def on_tray_icon_activated(self, reason):
-        if reason == QSystemTrayIcon.DoubleClick:
-            self.parent().show()
-            self.parent().activateWindow()
-
-    def add_character(self):
-        if len(self.character_widgets) < self.max_characters:
-            character = CharacterWidget(self.parent())
-            character.show()
-            character.exit_signal.connect(
-                lambda: self.remove_specific_character(character)
-            )
-            self.character_widgets.append(character)
-            self.remove_character_action.setEnabled(True)
-            if len(self.character_widgets) == self.max_characters:
-                self.add_character_action.setEnabled(False)
-        else:
-            self.showMessage(
-                "Ari",
-                f"최대 캐릭터 수({self.max_characters}마리)에 도달했습니다.",
-                QSystemTrayIcon.Information,
-                2000,
-            )
-
-    def remove_character(self):
-        if self.character_widgets:
-            character = self.character_widgets.pop()
-            character.close()
-            if not self.character_widgets:
-                self.remove_character_action.setEnabled(False)
-            self.add_character_action.setEnabled(True)
-        else:
-            self.showMessage(
-                "Ari",
-                "제거할 캐릭터가 없습니다.",
-                QSystemTrayIcon.Information,
-                2000,
-            )
-
-    def remove_specific_character(self, character):
-        if character in self.character_widgets:
-            self.character_widgets.remove(character)
-            character.close()
-            if not self.character_widgets:
-                self.remove_character_action.setEnabled(False)
-            self.add_character_action.setEnabled(True)
-            if hasattr(self.parent(), "update_character_count"):
-                self.parent().update_character_count()
-
-    def start_random_move(self):
-        for character in self.character_widgets:
-            character.start_random_move()
-
 
 class VoiceRecognitionThread(QThread):
     result = Signal(str)
@@ -1362,470 +653,6 @@ class VoiceRecognitionThread(QThread):
         self.init_microphone()
 
 
-class CharacterWidget(QWidget):
-    exit_signal = Signal()
-    toggle_voice_recognition = Signal()
-    set_listening_state_signal = Signal(bool)
-    show_speech_bubble_signal = Signal(str)
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.is_listening = False
-        self.setWindowFlags(Qt.Tool | Qt.FramelessWindowHint | Qt.WindowStaysOnTopHint)
-        self.setAttribute(Qt.WA_TranslucentBackground)
-        self.setAttribute(Qt.WA_ShowWithoutActivating)
-
-        self.offset = QPoint()
-        self.animation = None
-        self.context_menu = None
-        self.is_moving = False
-        self.is_dragging = False
-        self.current_frame = 0
-        self.last_pos = self.pos()
-
-        self.facing_left = False  # 기본적으로 오른쪽을 바라봄
-        self.actions = ["idle", "walk", "sit"]
-        self.current_action = "idle"
-
-        self.image_cache = {}
-        self.falling = False
-        self.fall_frame = 0
-
-        self.speech_bubble = None
-        self.speech_timer = QTimer(self)
-        self.speech_timer.timeout.connect(self.hide_speech_bubble)
-
-        # 폰트 로드
-        font_id = QFontDatabase.addApplicationFont("DNFBitBitv2.ttf")
-        if font_id != -1:
-            self.font_family = QFontDatabase.applicationFontFamilies(font_id)[0]
-        else:
-            logging.warning(
-                "DNFBitBitv2.ttf 폰트를 로드할 수 없습니다. 기본 폰트를 사용합니다."
-            )
-            self.font_family = QFont().family()
-
-        self.initUI()
-        self.load_images()
-        self.start_auto_move()
-
-        self.animation_thread = CharacterAnimationThread(self)
-        self.animation_thread.update_signal.connect(self.update_animation)
-        self.animation_thread.start()
-
-        self.show_speech_bubble_signal.connect(self._show_speech_bubble)
-
-    def initUI(self):
-        self.setGeometry(100, 100, 100, 100)
-
-        self.move_timer = QTimer(self)
-        self.move_timer.timeout.connect(self.start_random_move)
-
-        self.action_timer = QTimer(self)
-        self.action_timer.timeout.connect(self.perform_random_action)
-
-        self.action_duration_timer = QTimer(self)
-        self.action_duration_timer.timeout.connect(self.end_current_action)
-
-        self.fall_timer = QTimer(self)
-        self.fall_timer.timeout.connect(self.fall)
-
-        self.interaction_timer = QTimer(self)
-        self.interaction_timer.timeout.connect(self.interact_with_others)
-        self.interaction_timer.start(5000)  # 5초마다 상호작용 시도
-
-        self.set_listening_state_signal.connect(self._set_listening_state)
-        self.create_context_menu()
-
-    def create_context_menu(self):
-        self.context_menu = QMenu(self)
-        sit_action = self.context_menu.addAction("앉기")
-        sit_action.triggered.connect(self.sit)
-        idle_action = self.context_menu.addAction("기본 상태")
-        idle_action.triggered.connect(
-            self.return_to_idle
-        )  # idle() 대신 return_to_idle() 사용
-        fall_action = self.context_menu.addAction("떨어지기")
-        fall_action.triggered.connect(self.start_fall)
-        exit_action = self.context_menu.addAction("제거")
-        exit_action.triggered.connect(self.close)
-
-    def load_images(self):
-        image_folder = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)), "images"
-        )
-        self.image_sets = {
-            "idle": [
-                self.load_and_cache_image(os.path.join(image_folder, f"idle{i}.png"))
-                for i in range(1, 9)
-            ],
-            "walk": [
-                self.load_and_cache_image(os.path.join(image_folder, f"walk{i}.png"))
-                for i in range(1, 10)
-            ],
-            "drag": [
-                self.load_and_cache_image(os.path.join(image_folder, f"drag{i}.png"))
-                for i in range(1, 9)
-            ],
-            "listen": [
-                self.load_and_cache_image(os.path.join(image_folder, f"sit{i}.png"))
-                for i in range(1, 10)
-            ],
-            "sit": [
-                self.load_and_cache_image(os.path.join(image_folder, f"sit{i}.png"))
-                for i in range(1, 10)
-            ],
-            "fall": [
-                self.load_and_cache_image(os.path.join(image_folder, f"fall{i}.png"))
-                for i in range(1, 9)
-            ],
-        }
-        self.current_image = self.image_sets["idle"][0]
-        self.resize(QPixmap.fromImage(self.current_image).size())
-
-    def load_and_cache_image(self, path):
-        if path not in self.image_cache:
-            image = QImage(path)
-            scaled_image = image.scaled(
-                image.width() * 1.5,
-                image.height() * 1.5,
-                Qt.KeepAspectRatio,
-                Qt.SmoothTransformation,
-            )
-            self.image_cache[path] = scaled_image
-        return self.image_cache[path]
-
-    def start_auto_move(self):
-        if not self.is_listening:
-            self.move_timer.start(random.randint(3000, 10000))
-
-    def stop_auto_move(self):
-        self.move_timer.stop()
-        if self.animation:
-            self.animation.stop()
-        self.is_moving = False
-
-    def start_random_move(self):
-        if (
-            not self.is_dragging
-            and not self.is_listening
-            and self.current_action != "sit"
-            and not self.falling
-        ):
-            self.animate()
-
-    def animate(self):
-        if (
-            (
-                self.animation is None
-                or self.animation.state() == QPropertyAnimation.Stopped
-            )
-            and not self.is_listening
-            and self.current_action != "sit"
-            and not self.falling
-        ):
-            self.animation = QPropertyAnimation(self, b"pos")
-            self.animation.setDuration(3000)
-            start_pos = self.pos()
-            self.animation.setStartValue(start_pos)
-
-            screen = QApplication.primaryScreen().geometry()
-            max_distance = 100
-            new_x = max(
-                0,
-                min(
-                    start_pos.x() + random.randint(-max_distance, max_distance),
-                    screen.width() - self.width(),
-                ),
-            )
-            new_y = max(
-                0,
-                min(
-                    start_pos.y() + random.randint(-max_distance, max_distance),
-                    screen.height() - self.height(),
-                ),
-            )
-
-            self.animation.setEndValue(QPoint(new_x, new_y))
-            self.animation.setEasingCurve(QEasingCurve.InOutQuad)
-            self.animation.finished.connect(self.animationFinished)
-            self.animation.start()
-            self.current_action = "walk"
-            self.is_moving = True
-            self.facing_left = new_x < start_pos.x()
-
-    def animationFinished(self):
-        self.is_moving = False
-        self.current_action = "idle"
-        self.update()
-
-    def update_animation(self):
-        action = (
-            "fall"
-            if self.falling
-            else (
-                "drag"
-                if self.is_dragging
-                else (
-                    "walk"
-                    if self.is_moving and not self.is_listening
-                    else (
-                        "sit"
-                        if self.current_action == "sit" or self.is_listening
-                        else "idle"
-                    )
-                )
-            )
-        )
-        self.current_frame = (self.current_frame + 1) % len(self.image_sets[action])
-        self.current_image = self.image_sets[action][self.current_frame]
-        self.update()
-
-    def paintEvent(self, event):
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.Antialiasing)
-
-        if self.facing_left:
-            flipped_image = self.current_image.mirrored(True, False)
-            painter.drawPixmap(self.rect(), QPixmap.fromImage(flipped_image))
-        else:
-            painter.drawPixmap(self.rect(), QPixmap.fromImage(self.current_image))
-
-        if self.speech_bubble:
-            self.speech_bubble.update_position()
-
-    def mousePressEvent(self, event):
-        if event.button() == Qt.LeftButton:
-            self.offset = event.pos()
-            self.is_dragging = True
-            self.stop_auto_move()
-            if self.current_action == "sit":
-                self.return_to_idle()  # 앉아있는 상태에서 드래그 시 idle 상태로 변경
-            self.current_image = self.image_sets["drag"][0]
-            self.update()
-        elif event.button() == Qt.RightButton:
-            self.show_context_menu(event.pos())
-
-    def mouseMoveEvent(self, event):
-        if event.buttons() & Qt.LeftButton:
-            new_pos = self.mapToParent(event.pos() - self.offset)
-            self.move(new_pos)
-            self.update_drag_image(new_pos)
-
-    def mouseReleaseEvent(self, event):
-        if event.button() == Qt.LeftButton:
-            self.is_dragging = False
-            self.return_to_idle()  # 드래그 종료 시 idle 상태로 변경
-            self.start_auto_move()
-
-    def mouseDoubleClickEvent(self, event):
-        if event.button() == Qt.LeftButton:
-            self.start_fall()
-
-    def update_drag_image(self, new_pos):
-        if new_pos.x() < self.last_pos.x():
-            self.facing_left = True
-        elif new_pos.x() > self.last_pos.x():
-            self.facing_left = False
-        self.last_pos = new_pos
-
-    def show_context_menu(self, pos):
-        self.context_menu.exec_(self.mapToGlobal(pos))
-
-    def closeEvent(self, event):
-        self.exit_signal.emit()
-        super().closeEvent(event)
-
-    def sit(self):
-        self.stop_auto_move()
-        self.current_action = "sit"
-        self.current_image = self.image_sets["sit"][0]
-        self.update()
-        self.action_timer.stop()  # 앉아있는 동안 자동 행동 타이머 중지
-
-    def idle(self):
-        self.return_to_idle()
-        self.start_auto_move()
-
-    def return_to_idle(self):
-        self.current_action = "idle"
-        self.current_image = self.image_sets["idle"][0]
-        self.update()
-        self.action_timer.start(random.randint(15000, 45000))  # 자동 행동 타이머 재시작
-
-    def _set_listening_state(self, is_listening):
-        self.is_listening = is_listening
-        if is_listening:
-            self.current_action = "sit"
-            self.current_image = self.image_sets["sit"][0]
-            self.stop_auto_move()
-            self.action_timer.stop()
-            self.action_duration_timer.stop()
-        else:
-            self.current_action = "idle"
-            self.current_image = self.image_sets["idle"][0]
-            self.start_auto_move()
-            self.action_timer.start(random.randint(15000, 45000))
-        self.update()
-
-    def set_listening_state(self, is_listening):
-        self.set_listening_state_signal.emit(is_listening)
-
-    def perform_random_action(self):
-        if (
-            not self.is_listening
-            and not self.is_dragging
-            and self.current_action != "sit"
-            and not self.falling
-        ):
-            action = random.choice(["idle", "walk", "sit"])
-            if action == "walk":
-                self.start_random_move()
-            elif action == "sit":
-                self.sit()
-                duration = random.randint(3000, 8000)
-                self.action_duration_timer.start(duration)
-            else:
-                self.return_to_idle()  # idle() 대신 return_to_idle() 사용
-
-        if not self.is_listening and self.current_action != "sit" and not self.falling:
-            self.action_timer.start(random.randint(15000, 45000))
-
-    def end_current_action(self):
-        self.action_duration_timer.stop()
-        self.idle()
-
-    def start_fall(self):
-        if not self.falling and self.current_action != "sit":
-            self.falling = True
-            self.stop_auto_move()
-            self.fall_frame = 0
-            self.fall_timer.start(50)
-
-    def fall(self):
-        if not self.falling:
-            return
-
-        screen = QApplication.primaryScreen().geometry()
-        current_pos = self.pos()
-        new_pos = QPoint(current_pos.x(), current_pos.y() + 10)
-
-        if new_pos.y() + self.height() > screen.height():
-            self.falling = False
-            new_pos.setY(screen.height() - self.height())
-            self.current_action = "idle"  # 떨어진 후 idle 상태로 변경
-            self.start_auto_move()
-
-        self.move(new_pos)
-
-    def interact_with_others(self):
-        if self.parent() and hasattr(self.parent(), "character_widgets"):
-            others = [char for char in self.parent().character_widgets if char != self]
-            if others:
-                target = random.choice(others)
-                self.move_towards(target.pos())
-
-    def move_towards(self, target_pos):
-        if (
-            self.falling
-            or self.is_dragging
-            or self.is_listening
-            or self.current_action == "sit"
-        ):
-            return
-
-        current_pos = self.pos()
-        dx = target_pos.x() - current_pos.x()
-        dy = target_pos.y() - current_pos.y()
-        distance = (dx**2 + dy**2) ** 0.5
-
-        if distance > 5:
-            speed = 5
-            ratio = speed / distance
-            new_x = current_pos.x() + dx * ratio
-            new_y = current_pos.y() + dy * ratio
-            self.move(QPoint(int(new_x), int(new_y)))
-            self.current_frame = (self.current_frame + 1) % len(self.walk_images)
-            self.current_image = self.walk_images[self.current_frame]
-            self.facing_left = dx < 0
-            self.update()
-
-    @Slot(str)
-    def show_speech_bubble(self, text):
-        # 메인 스레드에서 실행되도록 시그널 발생
-        self.show_speech_bubble_signal.emit(text)
-
-    @Slot(str)
-    def _show_speech_bubble(self, text):
-        if self.speech_bubble:
-            self.speech_bubble.hide()
-
-        self.speech_bubble = SpeechBubble(text, self)
-        self.speech_bubble.show()
-
-    @Slot()
-    def hide_speech_bubble(self):
-        if self.speech_bubble:
-            self.speech_bubble.hide()
-            self.speech_bubble = None
-            self.speech_bubble = None
-
-
-class SpeechBubble(QWidget):
-    def __init__(self, text, parent):
-        super().__init__(parent)
-        self.text = text
-        self.parent = parent
-        self.setWindowFlags(Qt.ToolTip | Qt.FramelessWindowHint)
-        self.setAttribute(Qt.WA_TranslucentBackground)
-        self.font = QFont(parent.font_family, 12)
-        self.fm = QFontMetrics(self.font)
-        self.padding = 10
-        self.calculate_size()
-        self.update_position()
-
-    def calculate_size(self):
-        max_width = 300  # 최대 너비
-        text_width = self.fm.horizontalAdvance(self.text)
-        if text_width <= max_width:
-            # 한 줄로 표시 가능한 경우
-            self.bubble_width = text_width + self.padding * 2
-            self.bubble_height = self.fm.height() + self.padding * 2
-        else:
-            # 여러 줄로 표시해야 하는 경우
-            self.bubble_width = max_width
-            rect = self.fm.boundingRect(
-                QRect(0, 0, max_width, 1000), Qt.TextWordWrap, self.text
-            )
-            self.bubble_height = rect.height() + self.padding * 2
-
-        self.setFixedSize(self.bubble_width, self.bubble_height)
-
-    def paintEvent(self, event):
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.Antialiasing)
-
-        # 말풍선 그리기
-        painter.setPen(Qt.NoPen)
-        painter.setBrush(QColor(255, 255, 255, 200))
-        painter.drawRoundedRect(self.rect(), 10, 10)
-
-        # 텍스트 그리기
-        painter.setPen(Qt.black)
-        painter.setFont(self.font)
-        text_rect = self.rect().adjusted(
-            self.padding, self.padding, -self.padding, -self.padding
-        )
-        painter.drawText(text_rect, Qt.AlignCenter | Qt.TextWordWrap, self.text)
-
-    def update_position(self):
-        parent_rect = self.parent.rect()
-        parent_pos = self.parent.mapToGlobal(parent_rect.topLeft())
-        x = parent_pos.x() + (parent_rect.width() - self.bubble_width) // 2
-        y = parent_pos.y() - self.bubble_height - 10
-        self.move(x, y)
-
-
 # TTS 스레드
 class TTSThread(QThread):
     def __init__(self):
@@ -1860,189 +687,3 @@ class CommandExecutionThread(QThread):
 
     def execute(self, command):
         self.queue.put(command)
-
-
-# 캐릭터 애니메이션 스레드
-class CharacterAnimationThread(QThread):
-    update_signal = Signal()
-
-    def __init__(self, character_widget):
-        super().__init__()
-        self.character_widget = character_widget
-
-    def run(self):
-        while True:
-            time.sleep(0.1)  # 100ms마다 업데이트
-            self.update_signal.emit()
-
-
-class MainWindow(QMainWindow):
-    def __init__(self):
-        super().__init__()
-        self.setWindowTitle("Ari Voice Command")
-        self.setGeometry(100, 100, 300, 250)
-        self.setFixedSize(300, 200)
-
-        global icon_path
-        if icon_path:
-            self.tray_icon = SystemTrayIcon(QIcon(icon_path), self)
-        else:
-            self.tray_icon = SystemTrayIcon(QIcon(), self)  # 빈 아이콘으로 생성
-        self.tray_icon.show()
-
-        self.tray_icon.add_character()
-
-        self.initUI()
-
-        self.model_loading_thread = ModelLoadingThread()
-        self.model_loading_thread.finished.connect(self.on_model_loaded)
-        self.model_loading_thread.start()
-        self.tray_icon.showMessage(
-            "Ari", "TTS 및 Whisper 모델 로딩 중...", QSystemTrayIcon.Information, 3000
-        )
-
-        self.resource_monitor = ResourceMonitor()
-        self.resource_monitor.gc_needed.connect(perform_gc)
-        self.resource_monitor.start()
-
-        self.voice_thread = VoiceRecognitionThread()
-        self.tts_thread = TTSThread()
-        self.command_thread = CommandExecutionThread()
-
-        self.voice_thread.result.connect(self.handle_voice_result)
-        self.voice_thread.listening_state_changed.connect(self.update_listening_state)
-
-        self.voice_thread.start()
-        self.tts_thread.start()
-        self.command_thread.start()
-
-        # GC 타이머 추가
-        self.gc_timer = QTimer(self)
-        self.gc_timer.timeout.connect(self.periodic_gc)
-        self.gc_timer.start(5 * 60 * 1000)  # 5분마다 실행 (밀리초 단위)
-
-    def initUI(self):
-        self.microphone_label = QLabel("마이크 선택:", self)
-        self.microphone_label.setGeometry(20, 50, 80, 30)
-
-        self.microphone_combo = QComboBox(self)
-        self.microphone_combo.setGeometry(100, 50, 180, 30)
-        self.microphone_combo.addItems(sr.Microphone.list_microphone_names())
-
-        self.save_button = QPushButton("저장", self)
-        self.save_button.setGeometry(100, 100, 100, 30)
-        self.save_button.clicked.connect(self.save_settings)
-
-    def save_settings(self):
-        selected_microphone = self.microphone_combo.currentText()
-        logging.info(f"선택된 마이크: {selected_microphone}")
-        try:
-            if hasattr(self, "voice_thread") and self.voice_thread:
-                self.voice_thread.selected_microphone = selected_microphone
-                self.voice_thread.init_microphone()
-            self.hide()
-            self.tray_icon.showMessage(
-                "Ari",
-                "마이크 설정이 저장되었습니다.",
-                QSystemTrayIcon.Information,
-                2000,
-            )
-        except Exception as e:
-            logging.error(f"마이크 설정 저장 중 오류 발생: {str(e)}")
-            self.tray_icon.showMessage(
-                "Ari",
-                "마이크 설정 저장 중 오류가 발생했습니다.",
-                QSystemTrayIcon.Warning,
-                2000,
-            )
-
-    def closeEvent(self, event):
-        event.ignore()
-        self.hide()
-
-    def on_model_loaded(self):
-        logging.info("모델 로딩이 완료되었습니다.")
-        # 로딩 완료 알림
-        self.tray_icon.showMessage(
-            "Ari", "TTS 및 Whisper 모델 로딩 완료!", QSystemTrayIcon.Information, 3000
-        )
-        # 여기에 모델 로딩 완료 후 수행할 추가 작업을 넣을 수 있습니다.
-
-    def show_loading_progress(self, message):
-        self.tray_icon.showMessage("Ari", message, QSystemTrayIcon.Information, 2000)
-
-    def show_window(self):
-        self.show()
-        self.activateWindow()
-
-    def handle_voice_result(self, text):
-        logging.info(f"인식된 명령: {text}")
-        self.command_thread.execute(text)
-
-    def update_listening_state(self, is_listening):
-        for character in self.tray_icon.character_widgets:
-            character.set_listening_state(is_listening)
-
-    def periodic_gc(self):
-        logging.info("주기적 가비지 컬렉션 실행")
-        perform_gc()
-
-
-def main():
-    global ai_assistant, icon_path, tts_thread
-    setproctitle.setproctitle("Ari Voice Command")
-    try:
-        setup_logging()
-        logging.info("프로그램 시작")
-
-        app = QApplication(sys.argv)
-        main_window = MainWindow()
-        main_window.show()
-
-        if not QSystemTrayIcon.isSystemTrayAvailable():
-            logging.error("시스템 트레이를 사용할 수 없습니다.")
-            sys.exit(1)
-
-        app.setQuitOnLastWindowClosed(False)
-
-        icon_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "icon.png")
-        if not os.path.exists(icon_path):
-            logging.warning("아이콘 파일이 없습니다. 기본 아이콘을 사용합니다.")
-            icon_path = None
-
-        tray_icon = main_window.tray_icon
-        tts_thread = main_window.tts_thread
-
-        # AI 어시스턴트 초기화
-        try:
-            ai_assistant = get_ai_assistant()
-        except Exception as e:
-            logging.error(f"AI 어시스턴트 초기화 실패: {str(e)}")
-            sys.exit(1)
-
-        # 마이크 설정 변경 시 음성 인식 스레드에 알림
-        def update_microphone():
-            selected_microphone = main_window.microphone_combo.currentText()
-            main_window.voice_thread.set_microphone(selected_microphone)
-
-        main_window.save_button.clicked.connect(update_microphone)
-
-        # 종료 시 정리
-        app.aboutToQuit.connect(main_window.voice_thread.stop)
-        app.aboutToQuit.connect(main_window.voice_thread.wait)
-        app.aboutToQuit.connect(lambda: main_window.tts_thread.queue.put(None))
-        app.aboutToQuit.connect(lambda: main_window.command_thread.queue.put(None))
-        app.aboutToQuit.connect(main_window.tts_thread.wait)
-        app.aboutToQuit.connect(main_window.command_thread.wait)
-        app.aboutToQuit.connect(main_window.tray_icon.exit)
-        app.aboutToQuit.connect(main_window.resource_monitor.stop)
-        app.aboutToQuit.connect(main_window.resource_monitor.wait)
-
-        sys.exit(app.exec())
-    except Exception as e:
-        logging.error(f"예외 발생: {str(e)}", exc_info=True)
-        sys.exit(1)
-
-
-if __name__ == "__main__":
-    main()
